refactor: consolidate temp data_prep scripts into temperature_events.py
Merges heat_cold_bounds_case.py, plot_temperature_events.py, and
temperature_bounds_global.py into a single temperature_events.py with
three subcommands (plot / case / global). Preserves develop behavior
exactly: original _wrap_periodically formula, no ndimage landmass
isolation, numpy-based detect_events / expand_event_bounds /
include_temps_with_events, and no --event-quantiles flag.

Made-with: Cursor
